modularize the sort algorithms. properly handle `debug` and `help_text`
